pre-built some windows stuff

+++ include/tracking/Tracker.hpp
@@ -16,22 +16,22 @@ public:
   void addBuffer(unsigned char* buffer, int size);
   void Start();
   void Stop();
-  void define_image_counter(long int *buffer,int size);
+  void define_image_counter(int *buffer,int size);
   float get_x(int bdnb);
   float get_y(int bdnb);
   void set_saving(std::string file);
   void define_timebuffer(uint64_t *buffer);
   std::vector<float> get_copy_x_array(int i);
   std::vector<float> get_copy_y_array(int i);
-  void set_frames_to_analyze(long int a){*_frames_to_analyze=a;};
+  void set_frames_to_analyze(int a){*_frames_to_analyze=a;};
 
 
 
 
 
   int ci;
-  long int *_frames_to_analyze;
-  long int nb_frames_analyzed;
+  int *_frames_to_analyze;
+  int nb_frames_analyzed;
   int buffer_size;
   int beadtracking;
   O_i ** O_il;
